+++ components/ComponentLabel/index.js
@@ -1,8 +1,8 @@
 import styles from './styles.module.scss'
 
-export function ComponentLabel({ label, children }) {
+export function ComponentLabel({ label, children, className }) {
   return (
-    <div className={styles.container}>
+    <div className={`${styles.container} ${className}`}>
       <span className={styles.label}>{label}</span>
       <div className={styles.children}>{children}</div>
     </div>