+++ components/PostCard/index.js
@@ -2,8 +2,7 @@ import { ROUTER_PATH } from 'constants/routerPath';
 import Link from 'next/link';
 import styles from './styles.module.scss';
 
-export function PostCard(props) {
-  const { id, title, children = 'Click to view page details' } = props;
+export function PostCard({ id, title, children = 'Click to view page details' }) {
   return (
     <Link href={{
       pathname: ROUTER_PATH.DETAIL_POST,