+++ components/index.js
@@ -5,3 +5,7 @@ export * from './Card';
 export * from './Backdrop';
 export * from './PostCard';
 export * from './Breadcrumbs';
+export * from './HeaderAction';
+export * from './DataNotFound';
+export * from './ErrorMsg';
+export * from './Container';