+++ constants/routerPath.js
@@ -7,11 +7,13 @@ export const ROUTER_PATH = {
   ROUTER_1: '/router/router-1',
   POST: '/post',
   DETAIL_POST: '/post/detail-post',
+  CREATE_POST: '/post/create-post',
 }
 
 export const PARSE_ROUTER_PATH_BREADCRUMBS = {
+  'router-1': ROUTER_PATH.ROUTER_1,
   'detail-post': ROUTER_PATH.DETAIL_POST,
-  'router-1': ROUTER_PATH.ROUTER_1
+  'create-post': ROUTER_PATH.CREATE_POST,
 }
 
 export default {}